lib/output: migrate WriteMarkdown to main Output implementation
